D3D_4Q_1st - 파일명 수정 - 암묵적 형변환 제거 - 에러메시지 throw
D3D_4Q_2nd
- IBL 적용
- 배경 추가

D3D_Practice_9thWeek
- 코드 다듬기

+++ D3D_4Q_1st/DemoGameApp.cpp
@@ -597,8 +597,8 @@ void DemoGameApp::SetCube()
 	m_Camera.m_MoveSpeed = 500;
 
 	//fov 초기화
-	fovWidht = m_Width;
-	fovHeight = m_Height;
+	fovWidht = (float)m_Width;
+	fovHeight = (float)m_Height;
 	m_ProjectionMatrix = DirectX::XMMatrixPerspectiveFovLH(DirectX::XMConvertToRadians(FieldOfView), (float)m_Width / m_Height, m_near, m_far);
 
 	//fbx 파일 로드
